The sboot seal command now toggles the bootm command to seal before passing control to the OS. This allows for complex commands to run before boot.
Both bootm and SPL can enforce a correct unseal with the config options
CONFIG_SBOOT_ENFORCE and CONFIG_SPL_SBOOT_ENFORCE. In enforce mode, a
failure will cause the PCRs to be extended with a random value, and the
running application to hang.

U-Boot and the OS seal data is stored in TPM NVRAM. This require (a) the
TPM to support >= 624bytes of user defineable space (b) two indexes
defined with PPWRITE|PPREAD permissions and (c) the ability to software
assert physical presence. The data storage can be made more robust by
letting the user choose the permissions (for instances where software
assertion of physical presence is not supported).